requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and Language Processing
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20

# Google Calendar Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Date/Time Processing
python-dateutil==2.8.2
pytz==2023.3

# Data Validation and Serialization
pydantic==2.5.0

# HTTP Client
httpx==0.25.2

# Environment and Configuration
python-dotenv==1.0.0

# Security and Encryption
cryptography==41.0.7

# Utilities
pathlib2==2.3.7

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Streamlit for frontend (if needed)
streamlit==1.28.1

# Optional: Database support
sqlalchemy==2.0.23
alembic==1.12.1

# Optional: Redis for caching
redis==5.0.1

# Optional: Monitoring and logging
prometheus-client==0.19.0
